src/requirements.txt
aiogram_calendar=0.1.1